Use the manifest_tool to create the Docker build context
This change improves the stability of the Docker build context by
explicitly constructing it ourselves instead of letting Docker handle
that. Notably, this makes our files be immune to changes in the user's
umask (which makes the files have different permission bits, which
always invalidates the build cache) and mtime (which according to some
StackOverflow lore _may_ invalidate the build cache in some cases, but
I'm not quite positive about this).